schema: Move rotation and transform out of dev
