+++ common/schema/BUILD.bazel
@@ -15,7 +15,20 @@ drake_cc_package_library(
     name = "schema",
     visibility = ["//visibility:public"],
     deps = [
+        ":rotation",
         ":stochastic",
+        ":transform",
+    ],
+)
+
+drake_cc_library(
+    name = "rotation",
+    srcs = ["rotation.cc"],
+    hdrs = ["rotation.h"],
+    deps = [
+        ":stochastic",
+        "//common:name_value",
+        "//math:geometric_transform",
     ],
 )
 
@@ -32,12 +45,45 @@ drake_cc_library(
     ],
 )
 
+drake_cc_library(
+    name = "transform",
+    srcs = ["transform.cc"],
+    hdrs = ["transform.h"],
+    deps = [
+        ":rotation",
+        ":stochastic",
+        "//common:name_value",
+        "//math:geometric_transform",
+    ],
+)
+
 # === test/ ===
+
+drake_cc_googletest(
+    name = "rotation_test",
+    deps = [
+        ":rotation",
+        "//common/test_utilities:eigen_matrix_compare",
+        "//common/yaml:yaml_read_archive",
+        "//common/yaml:yaml_write_archive",
+    ],
+)
 
 drake_cc_googletest(
     name = "stochastic_test",
     deps = [
         ":stochastic",
+        "//common/test_utilities:eigen_matrix_compare",
+        "//common/test_utilities:expect_throws_message",
+        "//common/test_utilities:symbolic_test_util",
+        "//common/yaml:yaml_read_archive",
+    ],
+)
+
+drake_cc_googletest(
+    name = "transform_test",
+    deps = [
+        ":transform",
         "//common/test_utilities:eigen_matrix_compare",
         "//common/test_utilities:symbolic_test_util",
         "//common/yaml:yaml_read_archive",
