--- common/schema/dev/BUILD.bazel
@@ -1,59 +1,0 @@
-# -*- mode: python -*-
-# vi: set ft=python :
-
-load(
-    "@drake//tools/skylark:drake_cc.bzl",
-    "drake_cc_googletest",
-    "drake_cc_library",
-)
-load("//tools/lint:lint.bzl", "add_lint_tests")
-
-package(default_visibility = ["//visibility:private"])
-
-drake_cc_library(
-    name = "rotation",
-    srcs = ["rotation.cc"],
-    hdrs = ["rotation.h"],
-    deps = [
-        "//common:name_value",
-        "//common/schema:stochastic",
-        "//math:geometric_transform",
-    ],
-)
-
-drake_cc_library(
-    name = "transform",
-    srcs = ["transform.cc"],
-    hdrs = ["transform.h"],
-    visibility = ["//multibody/parsing/dev:__pkg__"],
-    deps = [
-        ":rotation",
-        "//common:name_value",
-        "//common/schema:stochastic",
-        "//math:geometric_transform",
-    ],
-)
-
-# === test/ ===
-
-drake_cc_googletest(
-    name = "rotation_test",
-    deps = [
-        ":rotation",
-        "//common/test_utilities:eigen_matrix_compare",
-        "//common/yaml:yaml_read_archive",
-        "//common/yaml:yaml_write_archive",
-    ],
-)
-
-drake_cc_googletest(
-    name = "transform_test",
-    deps = [
-        ":transform",
-        "//common/test_utilities:eigen_matrix_compare",
-        "//common/test_utilities:symbolic_test_util",
-        "//common/yaml:yaml_read_archive",
-    ],
-)
-
-add_lint_tests()
