+++ common/schema/rotation.h
